+++ NEWS.md
@@ -12,11 +12,15 @@
 ## Tool templates and `create()`
 
 * `create()` no longer generates `man/` directory - roxygen2 now does
-  this automatically.
+  this automatically. It also no longer generates an package-level doc
+  template, instead call `use_package_doc()`.
 
 * New `use_rcpp()` sets up a package to use Rcpp.
 
 * New `use_knitr()` sets up a package to use knitr for vignettes.
+
+* New `use_package_doc()` sets up an Roxygen template for package-level
+  docs.
 
 * New function `install_svn()` to install an R package from a subversion
   repository.